renamed nuget id to semver.core
