+++ Semver.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
-VisualStudioVersion = 14.0.25123.0
+VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{AC069663-5FE9-438A-A9F4-C5293D695D98}"
 	ProjectSection(SolutionItems) = preProject
@@ -9,9 +9,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		README.md = README.md
 	EndProjectSection
 EndProject
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Semver", "Semver.Core\Semver.xproj", "{7BEE814B-CFF1-42DD-94F4-C23F4318A45E}"
-EndProject
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Semver.Tests", "Semver.Core.Tests\Semver.Tests.xproj", "{52A5E9DD-8595-43FA-BA90-213ADA318962}"
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Semver.Tests", "Semver.Test\Semver.Tests.xproj", "{52A5E9DD-8595-43FA-BA90-213ADA318962}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -19,10 +17,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7BEE814B-CFF1-42DD-94F4-C23F4318A45E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7BEE814B-CFF1-42DD-94F4-C23F4318A45E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7BEE814B-CFF1-42DD-94F4-C23F4318A45E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7BEE814B-CFF1-42DD-94F4-C23F4318A45E}.Release|Any CPU.Build.0 = Release|Any CPU
 		{52A5E9DD-8595-43FA-BA90-213ADA318962}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{52A5E9DD-8595-43FA-BA90-213ADA318962}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{52A5E9DD-8595-43FA-BA90-213ADA318962}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -30,5 +24,8 @@ Global
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{52A5E9DD-8595-43FA-BA90-213ADA318962} = {AC069663-5FE9-438A-A9F4-C5293D695D98}
 	EndGlobalSection
 EndGlobal
