--- Semver/project.json
@@ -1,32 +1,0 @@
-﻿{
-  "title": "Semantic versioning",
-  
-  "authors": [
-    "Max Hauser"
-  ],
-  "packOptions": {
-    "licenseUrl": "http://max.mit-license.org/",
-    "tags": [
-      "semver",
-      "semantic",
-      "version"
-    ],
-    "owners": [
-      "Max Hauser"
-    ],
-    "projectUrl": "https://github.com/maxhauser/semver/"
-  },
-  "description": "A semver implementation in .Net based on the v2.0.0 of the spec found at http://semver.org/.",
-  "copyright": "Copyright 2016",
-  "version": "2.0.0-*",
-  "frameworks": {
-    "netstandard1": {
-      "imports": "dnxcore50",
-      "dependencies": {
-        "NETStandard.Library": "1.5.0-rc2-24027"
-      }
-    },
-    "net20" : {
-    }
-  }
-}
