+++ semver.core/Properties/AssemblyInfo.cs
