+++ semver.core/project.lock.json
@@ -1,0 +1,16 @@
+{
+  "locked": false,
+  "version": 2,
+  "targets": {
+    ".NETFramework,Version=v2.0": {},
+    ".NETStandard,Version=v1.0": {}
+  },
+  "libraries": {},
+  "projectFileDependencyGroups": {
+    "": [],
+    ".NETFramework,Version=v2.0": [],
+    ".NETStandard,Version=v1.0": []
+  },
+  "tools": {},
+  "projectFileToolGroups": {}
+}
