+++ semver.core/semver.xproj
